// quizonline-frontend/src/app/api/generated/model/connection-event-read.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LoginMethodEnumDto } from './login-method-enum';


export interface ConnectionEventReadDto { 
    readonly id: number;
    account_email?: string;
    readonly created_at: string;
    login_method?: LoginMethodEnumDto;
    ip?: string | null;
    user_agent?: string;
    browser?: string;
    os?: string;
    local_time?: string;
    browser_language?: string;
    timezone?: string;
    screen_width?: number | null;
    screen_height?: number | null;
    online?: boolean | null;
    country?: string;
    country_code?: string;
    city?: string;
    region?: string;
    latitude?: number | null;
    longitude?: number | null;
}