// quizonline-frontend/src/app/api/generated/model/paginated-connection-event-read-list.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConnectionEventReadDto } from './connection-event-read';


export interface PaginatedConnectionEventReadListDto { 
    count: number;
    next?: string | null;
    previous?: string | null;
    results: Array<ConnectionEventReadDto>;
}